+++ src/lib/Components/Bidding/Components/BiddingThemeProvider.tsx
@@ -2,4 +2,8 @@ import React from "react"
 import { ThemeProvider } from "styled-components"
 import { theme } from "../Elements/Theme"
 
-export const BiddingThemeProvider = props => <ThemeProvider theme={theme} {...props} />
+export class BiddingThemeProvider extends React.Component {
+  render() {
+    return <ThemeProvider theme={theme} {...this.props} />
+  }
+}