+++ CHANGELOG.md
@@ -43,6 +43,7 @@
 * Fixes crash when following an artist from the ‘related artist’ Home/ForYou rail - alloy
 * Creates and verifies bidder position - sepans
 * Adds `<MarkdownRenderer>` component - sepans
+* Adds a `<Input>` component - yuki24
 
 ### 1.4.6